requirements.txt
# Core Dependencies
openai>=1.10.0
python-dotenv>=1.0.0

# Marcone API (SOAP)
zeep>=4.2.1
lxml>=4.9.0

# API & Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
requests>=2.31.0

# Salesforce Integration
simple-salesforce>=1.12.5

# Data Processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.1.0
